Refactor KtorHttpUtils to use top-level extensions
- Create ContentTypeUtils.kt with ContentType extensions (isText, isXml)
  and String.toContentTypeOrNull()
- Add String.toUrlOrNull() extension to UrlUtils.kt
- Rename KtorHttpUtils.kt to HttpUtils.kt and object to HttpUtils
- Update all usages of KtorHttpUtils to HttpUtils in DavResource.kt,
  Response.kt, ResponseParser.kt, PropStatParser.kt
- Create ContentTypeUtilsTest.kt with tests for ContentType extensions
- Rename KtorHttpUtilsTest.kt to HttpUtilsTest.kt
- Add tests for toUrlOrNull to UrlUtilsTest.kt
- Update deprecation in OkHttpUtils.kt to point to HttpUtils

This follows the same pattern as the UrlUtils.kt refactor, using top-level
extension methods where appropriate and grouping related utilities into
semantic files.
