+++ src/main/kotlin/at/bitfire/dav4jvm/ktor/UrlUtils.kt
@@ -128,3 +128,25 @@ fun Url.equalsForWebDAV(other: Url): Boolean {
             && this.port == other.port
             && this.rawSegments == other.rawSegments
 }
+
+/**
+ * Converts the [String] into a Ktor [Url], if possible.
+ *
+ * Differs from [io.ktor.http.parseUrl]:
+ *
+ * - `"relative".toUrlOrNull() == Url("relative")` (without host) but
+ * - `parseUrl("relative") == null` (requires host)
+ *
+ * @return the Ktor [Url], or `null` if the string couldn't be converted
+ */
+fun String?.toUrlOrNull(): Url? {
+    if (this == null)
+        return null
+
+    return try {
+        Url(this)
+    } catch (_: Exception) {
+        // parseUrl doesn't catch URLDecodeException, see https://github.com/ktorio/ktor/pull/5231
+        null
+    }
+}
