+++ src/main/kotlin/at/bitfire/dav4jvm/okhttp/OkHttpUtils.kt
@@ -13,7 +13,7 @@ package at.bitfire.dav4jvm.okhttp
 import okhttp3.HttpUrl
 import okhttp3.Response
 
-@Deprecated("Migrate to Ktor", ReplaceWith("at.bitfire.dav4jvm.ktor.KtorHttpUtils"))
+@Deprecated("Migrate to Ktor", ReplaceWith("at.bitfire.dav4jvm.ktor.HttpUtils"))
 object OkHttpUtils {
 
     /**
